Modify informer generators to have List functions parse more liberally (#1)
For CRDs that set preserveUnknownFields to true, informers can fail to sync because there is a mismatch between how strictly a resource is parsed between being put in kubernetes and how strictly a resource needs to be parsed before being put in a informer cache. This PR changes the List function to handle objects that don't parse properly differently by not including them in the returned list.